+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.2.1
 * change: interceptors.response onFulfilled function parameter (data, config, response) => (data, {url, config}, response)
+* update: interceptors preset
 ## 1.2.0
 * add: interceptors.use
 * add: interceptors preset abortRequest